Защитил Js, добавил состояния активным элементам, проверил валидатор и БЭМ дерево
