+++ source/index.html
@@ -54,13 +54,17 @@
         </div>
         <picture>
           <source type="image/webp" srcset="./img/jpg/main-screen-superman.webp 1x, ./img/jpg/main-screen-superman-retina.webp 2x">
-          <img scr="./img/jpg/main-screen-superman.jpg" srcset="./img/jpg/main-screen-superman-retina.jpg 2x" width="500" height="556" alt="Накаченный супермен.">
+          <img src="./img/jpg/main-screen-superman.jpg" srcset="./img/jpg/main-screen-superman-retina.jpg 2x" width="500" height="556" alt="Накаченный супермен.">
         </picture>
       </section>
     </div>
   </header>
+
+  <!-- Main -->
   <main class="grid-container">
     <h1 class="visually-hidden">Элитный фитнес-центр SUPERGYM в Омске</h1>
+
+    <!-- About as -->
     <section class="about-as">
       <div class="container about-as__container">
         <div class="about-as__text-wrapper">
@@ -80,7 +84,7 @@
         <div class="about-as__image-wrapper">
           <picture>
             <source type="image/webp" srcset="./img/jpg/about-as-gym.webp 1x, ./img/jpg/about-as-gym-retina.webp 2x">
-            <img scr="./img/jpg/about-as-gym.jpg" srcset="./img/jpg/about-as-gym-retina.jpg 2x" width="660" height="401" alt="Наш просторный тренажёрный зал.">
+            <img src="./img/jpg/about-as-gym.jpg" srcset="./img/jpg/about-as-gym-retina.jpg 2x" width="660" height="401" alt="Наш просторный тренажёрный зал.">
           </picture>
           <div class="about-as__video-wrapper about-as__video-mask" data-open-modal="video">
             <button class="about-as__play-button" type="button">
@@ -123,7 +127,7 @@
             <span>с 8:00 до 17:00</span>
             <b>
               1700
-              <span>7000</span>
+              <span>1700</span>
             </b>
             <a class="tub-list__link btn btn--by-subscriber" href="#" target="_blank">купить абонемент</a>
           </li>
@@ -152,7 +156,7 @@
             <span>с 8:00 до 17:00</span>
             <b>
               22700
-              <span>12700</span>
+              <span>22700</span>
             </b>
             <a class="tub-list__link btn btn--by-subscriber" href="#" target="_blank">купить абонемент</a>
           </li>
@@ -161,7 +165,7 @@
             <span>с 8:00 до 22:00 </span>
             <b>
               42700
-              <span>22700</span>
+              <span>42700</span>
             </b>
             <a class="tub-list__link btn btn--by-subscriber" href="#" target="_blank">купить абонемент</a>
           </li>
@@ -172,7 +176,7 @@
             <span>12 занятий</span>
             <b>
               32700
-              <span>22700</span>
+              <span>32700</span>
             </b>
             <a class="tub-list__link btn btn--by-subscriber" href="#" target="_blank">купить абонемент</a>
           </li>
@@ -181,7 +185,7 @@
             <span>с 8:00 до 17:00</span>
             <b>
               52700
-              <span>22700</span>
+              <span>52700</span>
             </b>
             <a class="tub-list__link btn btn--by-subscriber" href="#" target="_blank">купить абонемент</a>
           </li>
@@ -190,7 +194,7 @@
             <span>с 8:00 до 22:00 </span>
             <b>
               92700
-              <span>22700</span>
+              <span>92700</span>
             </b>
             <a class="tub-list__link btn btn--by-subscriber" href="#" target="_blank">купить абонемент</a>
           </li>
@@ -225,7 +229,7 @@
         </div>
         <picture>
           <source type="image/webp" srcset="./img/jpg/tournament.webp 1x, ./img/jpg/tournament-retina.webp 2x">
-          <img scr="./img/jpg/tournament.jpg" srcset="./img/jpg/tournament-retina.jpg 2x" width="660" height="500" alt="Борец на татами.">
+          <img src="./img/jpg/tournament.jpg" srcset="./img/jpg/tournament-retina.jpg 2x" width="660" height="500" alt="Борец на татами.">
         </picture>
       </div>
     </section>
@@ -660,9 +664,9 @@
           title="Бесплатные интерактивные онлайн-курсы"
           allow="accelerometer; clipboard-write; encrypted-media; gyroscope; picture-in-picture; autoplay"
           allowfullscreen></iframe>
-        <button class="modal__close-btn btn-reset" type="button" aria-label="Закрыть попап" data-close-modal>
+        <button class="modal__close-btn btn-reset" type="button" aria-label="Закрыть попап" data-close-modal data-stop-video>
           <svg width="40" height="40" viewBox="0 0 15 15" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-            <path d="M1 1L7.5 7.5M14 14L7.5 7.5M7.5 7.5L14 1L1 14" stroke-width="1" />
+            <use xlink:href="./img/sprite.svg#close-icon"></use>
           </svg>
         </button>
       </div>
