+++ source/sass/blocks/contacts.scss
@@ -49,6 +49,10 @@
     display: flex;
     align-items: center;
     color: #102654;
+
+    &:hover {
+      opacity: 0.7;
+    }
   }
 }
