+++ source/sass/blocks/footer-navigation.scss
@@ -25,6 +25,10 @@
   align-items: center;
   color: #ffffff;
   margin-bottom: 10px;
+
+  &:hover {
+    opacity: 0.7;
+  }
 }
 
 @media (max-width: 1199px) {
