+++ source/sass/blocks/footer-social.scss
@@ -25,6 +25,10 @@
   display: flex;
   align-items: center;
   justify-content: center;
+
+  &:hover {
+    opacity: 0.7;
+  }
 }
 
 @media (max-width: 1199px) {
