+++ source/sass/blocks/free-lesson-form.scss
@@ -40,6 +40,11 @@
     border: 3px solid #ffffff;
     width: 100%;
     min-height: 52px;
+    cursor: pointer;
+
+    &:hover {
+      opacity: 0.7;
+    }
   }
 }
