+++ source/sass/blocks/navigation-list.scss
@@ -23,6 +23,10 @@
   justify-content: center;
   align-items: center;
   padding: 4px 0;
+
+  &:hover {
+    opacity: 0.7;
+  }
 }
 
 @media (max-width: 1199px) {
